requirements.txt
django-filter>=2.1.0, <3.0
Markdown==3.2.1
jsonfield==2.1.0
Django==3.2.11
djangorestframework==3.12.4